Move websocket and api logic to their own files.
